feat: implement database CRUD query functions for users, products, and comments

+++ backend/src/db/queries.ts
@@ -14,16 +14,23 @@ export const getUserById = async (id: string) => {
 }
 
 export const updateUser = async (id: string, data: Partial<NewUser>) => {
-    const [updatedUser] = await db.update(users).set(data).where(eq(users.id, id)).returning();
-    return updatedUser;
+    const exsitingUser = await getUserById(id);
+    if (!exsitingUser) {
+        throw new Error("User not found");
+    } else {
+        const [updatedUser] = await db.update(users).set(data).where(eq(users.id, id)).returning();
+        return updatedUser;
+    }
+
 }
 
 export const upsertUser = async (data: NewUser) => {
-    const existingUser = await getUserById(data.id);
-
-    if (existingUser) return updateUser(data.id, data);
-
-    return createUser(data);
+    const [user] = await db.insert(users).values(data).onConflictDoUpdate({
+        target: users.id,
+        set: data,
+    })
+        .returning();
+    return user;
 }
 
 //PRODUCT QUERIES
@@ -66,13 +73,25 @@ export const getProductsByUserId = async (userId: string) => {
 }
 
 export const updateProduct = async (id: string, data: Partial<NewProduct>) => {
-    const [updatedProduct] = await db.update(products).set(data).where(eq(products.id, id)).returning();
-    return updatedProduct;
+    const exsitingProduct = await getProductById(id);
+    if (!exsitingProduct) {
+        throw new Error("Product not found");
+    } else {
+        const [updatedProduct] = await db.update(products).set(data).where(eq(products.id, id)).returning();
+        return updatedProduct;
+    }
 }
 
+
+
 export const deleteProduct = async (id: string) => {
-    const deletedProduct = await db.delete(products).where(eq(products.id, id)).returning();
-    return deletedProduct;
+    const exsitingProduct = await getProductById(id);
+    if (!exsitingProduct) {
+        throw new Error("Product not found");
+    } else {
+        const deletedProduct = await db.delete(products).where(eq(products.id, id)).returning();
+        return deletedProduct;
+    }
 }
 
 //COMMENT QUERIES
@@ -83,8 +102,13 @@ export const createComment = async (data: NewComment) => {
 }
 
 export const deleteComment = async (id: string) => {
-    const deleteComment = await db.delete(comments).where(eq(comments.id, id)).returning();
-    return deleteComment;
+    const exsitingComment = await getCommentById(id);
+    if (!exsitingComment) {
+        throw new Error("Comment not found");
+    } else {
+        const deleteComment = await db.delete(comments).where(eq(comments.id, id)).returning();
+        return deleteComment;
+    }
 }
 
 export const getCommentById = async (id: string) => {
